feat: 삭제 기능 UX개선

+++ vue-til/vue-til/src/components/posts/PostListItem.vue
@@ -26,8 +26,10 @@ export default {
   },
   methods: {
     async deleteItem() {
-      await deletePost(this.postItem._id);
-      console.log("deleted");
+      if (confirm("You want to delete it?")) {
+        await deletePost(this.postItem._id);
+        this.$emit("refresh");
+      }
     },
   },
 };
